+++ common/src/main/java/traben/entity_texture_features/config/ETFConfig.java
@@ -56,7 +56,6 @@ public class ETFConfig {
     public boolean removePixelsUnderEmissiveMobs = true;
     public boolean removePixelsUnderEmissiveBlockEntity = true;
 
-    public boolean zombiePiglinRightEarEnabled = false;
 
 
     public Set<ETFConfigScreenWarnings.ConfigWarning> ignoredConfigs = new HashSet<>();