+++ common/src/main/java/traben/entity_texture_features/mixin/entity/block_entity/MixinMobSpawnerBlockEntityRenderer.java
@@ -1,4 +1,4 @@
-package traben.entity_texture_features.mixin.entity.misc;
+package traben.entity_texture_features.mixin.entity.block_entity;
 
 import net.minecraft.block.entity.BellBlockEntity;
 import net.minecraft.client.render.block.entity.BlockEntityRenderer;