+++ common/src/main/java/traben/entity_texture_features/mixin/entity/renderer/feature/MixinCapeFeatureRenderer.java
@@ -1,4 +1,4 @@
-package traben.entity_texture_features.mixin.entity.feature_renderers;
+package traben.entity_texture_features.mixin.entity.renderer.feature;
 
 import net.minecraft.client.network.AbstractClientPlayerEntity;
 import net.minecraft.client.render.*;