+++ common/src/main/java/traben/entity_texture_features/texture_handlers/ETFManager.java
@@ -3,8 +3,9 @@ package traben.entity_texture_features.texture_handlers;
 import it.unimi.dsi.fastutil.objects.*;
 import net.minecraft.client.MinecraftClient;
 import net.minecraft.entity.Entity;
-import net.minecraft.entity.LivingEntity;
+import net.minecraft.entity.EntityType;
 import net.minecraft.entity.effect.StatusEffect;
+import net.minecraft.entity.mob.ZombifiedPiglinEntity;
 import net.minecraft.entity.passive.PandaEntity;
 import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.resource.Resource;
@@ -29,7 +30,7 @@ import static traben.entity_texture_features.ETFClientCommon.ETFConfigData;
 
 // ETF re-write
 //this class will ideally be where everything in vanilla interacts to get ETF stuff done
-public class ETFManager {
+public class ETFManager{
 
     public static final UUID ETF_GENERIC_UUID = UUID.nameUUIDFromBytes(("GENERIC").getBytes());
     private static final ETFTexture ETF_ERROR_TEXTURE = getErrorETFTexture();
@@ -69,6 +70,10 @@ public class ETFManager {
     private final Object2BooleanOpenHashMap<UUID> ENTITY_IS_UPDATABLE = new Object2BooleanOpenHashMap<>();
     private final ObjectOpenHashSet<UUID> ENTITY_UPDATE_QUEUE = new ObjectOpenHashSet<>();
     private final ObjectOpenHashSet<UUID> ENTITY_DEBUG_QUEUE = new ObjectOpenHashSet<>();
+
+    public final Object2IntOpenHashMap<EntityType<?>> ENTITY_TYPE_VANILLA_BRIGHTNESS_OVERRIDE_VALUE = new Object2IntOpenHashMap<>();
+
+    public final ObjectOpenHashSet<EntityType<?>> ENTITY_TYPE_IGNORE_PARTICLES = new ObjectOpenHashSet<>();
     //contains the total number of variants for any given vanilla texture
     private final Object2IntOpenHashMap<Identifier> TRUE_RANDOM_COUNT_CACHE = new Object2IntOpenHashMap<>();
     private final Object2LongOpenHashMap<UUID> LAST_PLAYER_CHECK_TIME = new Object2LongOpenHashMap<>();
@@ -80,6 +85,8 @@ public class ETFManager {
     public Boolean lecternHasCustomTexture = null;
     public ETFTexture redMooshroomAlt = null;
     public ETFTexture brownMooshroomAlt = null;
+
+    public static boolean zombiePiglinRightEarEnabled = false;
 
     private ETFManager() {
 
@@ -123,6 +130,7 @@ public class ETFManager {
             ETFUtils2.logError("emissive suffixes could not be read: default emissive suffix '_e' used");
             EMISSIVE_SUFFIX_LIST.add("_e");
         }
+        ENTITY_TYPE_VANILLA_BRIGHTNESS_OVERRIDE_VALUE.defaultReturnValue(0);
     }
 
     public static ETFManager getInstance() {
@@ -345,7 +353,7 @@ public class ETFManager {
                 }
             } else if (/*only*/possible2PNG == null) {
                 //optifine random confirmed
-                newOptifineTextureFound(vanillaIdentifier, possibleProperty);
+                newOptifineTextureFound(entity,vanillaIdentifier, possibleProperty);
                 return returnNewAlreadyConfirmedOptifineTexture(entity, vanillaIdentifier, false);
             } else //noinspection CommentedOutCode
             {//neither null this will be annoying
@@ -361,7 +369,7 @@ public class ETFManager {
 
                 // System.out.println("debug6534="+p2pngPackName+","+propertiesPackName+","+ETFUtils2.returnNameOfHighestPackFrom(packs));
                 if (propertiesPackName != null && propertiesPackName.equals(ETFUtils2.returnNameOfHighestPackFromTheseTwo(new String[]{p2pngPackName, propertiesPackName}))) {
-                    newOptifineTextureFound(vanillaIdentifier, possibleProperty);
+                    newOptifineTextureFound(entity,vanillaIdentifier, possibleProperty);
                     return returnNewAlreadyConfirmedOptifineTexture(entity, vanillaIdentifier, false);
                 } else {
                     if (source != TextureSource.ENTITY_FEATURE) {
@@ -377,12 +385,35 @@ public class ETFManager {
         return null;
     }
 
-    private void newOptifineTextureFound(Identifier vanillaIdentifier, Identifier properties) {
+    private void newOptifineTextureFound(Entity entity,Identifier vanillaIdentifier, Identifier properties) {
 
         try {
             Properties props = ETFUtils2.readAndReturnPropertiesElseNull(properties);
 
             if (props != null) {
+                //check for etf entity properties
+                if(props.containsKey("vanillaBrightnessOverride")){
+                    String value = props.getProperty("vanillaBrightnessOverride").trim();
+                    int tryNumber;
+                    try {
+                        tryNumber =Integer.parseInt(value.replaceAll("\\D", ""));
+                    }catch(NumberFormatException e){
+                        tryNumber = 0;
+                    }
+                    if(tryNumber >= 16) tryNumber = 15;
+                    if(tryNumber < 0) tryNumber = 0;
+                    ENTITY_TYPE_VANILLA_BRIGHTNESS_OVERRIDE_VALUE.put(entity.getType(),tryNumber);
+                }
+                if(entity instanceof ZombifiedPiglinEntity
+                        && props.containsKey("showHiddenModelParts")
+                        && props.getProperty("showHiddenModelParts").equals("true")) {
+                    zombiePiglinRightEarEnabled = true;
+                }
+                if(props.containsKey("suppressParticles")
+                        && props.getProperty("suppressParticles").equals("true")) {
+                    ENTITY_TYPE_IGNORE_PARTICLES.add(entity.getType());
+                }
+
                 Set<String> propIds = props.stringPropertyNames();
                 //set so only 1 of each
                 Set<Integer> numbers = new HashSet<>();
@@ -428,7 +459,7 @@ public class ETFManager {
                     @Nullable String[] maxHealthStrings = null;
                     @Nullable Integer[] inventoryColumns = null;
 //                    @Nullable Boolean isTrapHorse = null;
-//                    @Nullable Boolean isAngry = null;
+                   @Nullable Boolean isAngry = null;
                     @Nullable PandaEntity.Gene[] hiddenGene = null;
 //                    @Nullable Angriness[] wardenAngriness = null;
 //                    @Nullable Boolean isAngryWithClient = null;
@@ -746,14 +777,14 @@ public class ETFManager {
 //                            ETFUtils2.logWarn("properties files number error in trapHorse category");
 //                        }
 //                    }
-//                    if (props.containsKey("angry." + num)) {
-//                        String input = props.getProperty("angry." + num).trim();
-//                        if (input.equals("true") || input.equals("false")) {
-//                            isAngry = input.equals("true");
-//                        } else {
-//                            ETFUtils2.logWarn("properties files number error in angry category");
-//                        }
-//                    }
+                    if (props.containsKey("angry." + num)) {
+                        String input = props.getProperty("angry." + num).trim();
+                        if (input.equals("true") || input.equals("false")) {
+                            isAngry = input.equals("true");
+                        } else {
+                            ETFUtils2.logWarn("properties files number error in angry category");
+                        }
+                    }
                     if (props.containsKey("hiddenGene." + num)) {
                         String[] input = props.getProperty("hiddenGene." + num).trim().split("\s+");
                         ArrayList<PandaEntity.Gene> genes = new ArrayList<>();
@@ -900,7 +931,7 @@ public class ETFManager {
                                 maxHealthStrings,
                                 inventoryColumns,
 //                                isTrapHorse,
-//                                isAngry,
+                                isAngry,
                                 hiddenGene,
 //                                wardenAngriness,
 //                                isAngryWithClient,
@@ -990,7 +1021,7 @@ public class ETFManager {
         try {
             for (ETFTexturePropertyCase property :
                     optifineProperties) {
-                if (property.doesEntityMeetConditionsOfThisCase((LivingEntity) entity, isThisAnUpdate, ENTITY_IS_UPDATABLE)) {
+                if (property.doesEntityMeetConditionsOfThisCase(entity, isThisAnUpdate, ENTITY_IS_UPDATABLE)) {
                     return property.getAnEntityVariantSuffixFromThisCase(entity.getUuid());
                 }
             }