+++ common/src/main/java/traben/entity_texture_features/texture_handlers/ETFTexturePropertyCase.java
@@ -4,8 +4,8 @@ package traben.entity_texture_features.texture_handlers;
 import it.unimi.dsi.fastutil.objects.Object2BooleanOpenHashMap;
 import it.unimi.dsi.fastutil.objects.ObjectImmutableList;
 import net.minecraft.client.MinecraftClient;
+import net.minecraft.entity.Entity;
 import net.minecraft.entity.LivingEntity;
-import net.minecraft.entity.boss.WitherEntity;
 import net.minecraft.entity.effect.StatusEffect;
 import net.minecraft.entity.mob.*;
 import net.minecraft.entity.passive.*;
@@ -14,7 +14,6 @@ import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 import traben.entity_texture_features.ETFVersionDifferenceHandler;
 import traben.entity_texture_features.mixin.accessor.MooshroomEntityAccessor;
-import traben.entity_texture_features.utils.ETFCacheKey;
 import traben.entity_texture_features.utils.ETFUtils2;
 
 import java.util.ArrayList;
@@ -46,7 +45,7 @@ public class ETFTexturePropertyCase {
     private final @Nullable String[] MAX_HEALTH_STRINGS;
     private final @Nullable Integer[] INVENTORY_COLUMNS;
 //    private final @Nullable Boolean IS_TRAP_HORSE;
-//    private final @Nullable Boolean IS_ANGRY;
+    private final @Nullable Boolean IS_ANGRY;
     private final @Nullable PandaEntity.Gene[] HIDDEN_GENE;
 //    private final @Nullable Angriness[] WARDEN_ANGRINESS;
 //    private final @Nullable Boolean IS_ANGRY_WITH_CLIENT;
@@ -80,7 +79,7 @@ public class ETFTexturePropertyCase {
                                   @Nullable String[] maxHealthStrings,
                                   @Nullable Integer[] inventoryColumns,
 //                                  @Nullable Boolean isTrapHorse,
-//                                  @Nullable Boolean isAngry,
+                                 @Nullable Boolean isAngry,
                                   @Nullable PandaEntity.Gene[] hiddenGene,
 //                                  @Nullable Angriness[] wardenAngriness,
 //                                  @Nullable Boolean isAngryWithClient,
@@ -103,7 +102,7 @@ public class ETFTexturePropertyCase {
         MAX_HEALTH_STRINGS = maxHealthStrings;
         INVENTORY_COLUMNS = inventoryColumns;
 //        IS_TRAP_HORSE = isTrapHorse;
-//        IS_ANGRY = isAngry;
+        IS_ANGRY = isAngry;
         HIDDEN_GENE = hiddenGene;
 //        WARDEN_ANGRINESS = wardenAngriness;
 //        IS_ANGRY_WITH_CLIENT = isAngryWithClient;
@@ -170,7 +169,7 @@ public class ETFTexturePropertyCase {
     }
 
 
-    public boolean doesEntityMeetConditionsOfThisCase(LivingEntity entity, boolean isUpdate, Object2BooleanOpenHashMap<UUID> UUID_CaseHasUpdateablesCustom) {
+    public boolean doesEntityMeetConditionsOfThisCase(Entity entity, boolean isUpdate, Object2BooleanOpenHashMap<UUID> UUID_CaseHasUpdateablesCustom) {
 
 
         //System.out.println("checking property number "+propertyNumber);
@@ -194,7 +193,7 @@ public class ETFTexturePropertyCase {
                 && MAX_HEALTH_STRINGS == null
                 && INVENTORY_COLUMNS == null
 //                && IS_TRAP_HORSE == null
-//                && IS_ANGRY == null
+                && IS_ANGRY == null
                 && HIDDEN_GENE == null
 //                && WARDEN_ANGRINESS == null
 //                && IS_ANGRY_WITH_CLIENT == null
@@ -445,9 +444,9 @@ public class ETFTexturePropertyCase {
             }
             doesEntityMeetThisCaseTest = check;
         }
-        if (doesEntityMeetThisCaseTest && IS_BABY != 0) {
+        if (doesEntityMeetThisCaseTest && entity instanceof LivingEntity && IS_BABY != 0) {
             wasEntityTestedByAnUpdatableProperty = true;
-            doesEntityMeetThisCaseTest = (IS_BABY == 1) == entity.isBaby();
+            doesEntityMeetThisCaseTest = (IS_BABY == 1) == ((LivingEntity)entity).isBaby();
             //System.out.println("baby " + doesEntityMeetThisCaseTest);
         }
         if (doesEntityMeetThisCaseTest && WEATHER_TYPE != 0) {
@@ -472,12 +471,12 @@ public class ETFTexturePropertyCase {
             }
             doesEntityMeetThisCaseTest = check;
         }
-        if (doesEntityMeetThisCaseTest && HEALTH_RANGE_STRINGS != null) {
+        if (doesEntityMeetThisCaseTest && entity instanceof LivingEntity && HEALTH_RANGE_STRINGS != null) {
             wasEntityTestedByAnUpdatableProperty = true;
             //float entityHealth = entity.getHealth();
             boolean check = false;
             //always check percentage
-            float checkValue = entity.getHealth() / entity.getMaxHealth() * 100;
+            float checkValue = ((LivingEntity)entity).getHealth() / ((LivingEntity)entity).getMaxHealth() * 100;
             for (String hlth :
                     HEALTH_RANGE_STRINGS) {
                 if (hlth != null) {
@@ -666,8 +665,8 @@ public class ETFTexturePropertyCase {
         }
 
 
-        if (doesEntityMeetThisCaseTest && SPEED_MIN_MAX != null) {
-            double speed = entity.getMovementSpeed();
+        if (doesEntityMeetThisCaseTest && entity instanceof LivingEntity && SPEED_MIN_MAX != null) {
+            double speed = ((LivingEntity)entity).getMovementSpeed();
             Double min = SPEED_MIN_MAX[0];
             Double max = SPEED_MIN_MAX[1];
             if (min != null && max != null) {
@@ -682,10 +681,10 @@ public class ETFTexturePropertyCase {
                 doesEntityMeetThisCaseTest = (jumpHeight >= min && jumpHeight <= max);
             }
         }
-        if (doesEntityMeetThisCaseTest && MAX_HEALTH_STRINGS != null) {
+        if (doesEntityMeetThisCaseTest && entity instanceof LivingEntity && MAX_HEALTH_STRINGS != null) {
             boolean check = false;
             //always check percentage
-            float checkValue = entity.getMaxHealth();
+            float checkValue = ((LivingEntity)entity).getMaxHealth();
             for (String hlth :
                     MAX_HEALTH_STRINGS) {
                 if (hlth != null) {
@@ -723,18 +722,27 @@ public class ETFTexturePropertyCase {
 //            wasEntityTestedByAnUpdatableProperty = true;
 //            doesEntityMeetThisCaseTest = ((SkeletonHorseEntity) entity).isTrapped() == IS_TRAP_HORSE;
 //        }
-//        if (doesEntityMeetThisCaseTest && IS_ANGRY != null) {
-//            if (entity instanceof Angerable) {
-//                wasEntityTestedByAnUpdatableProperty = true;
-//                doesEntityMeetThisCaseTest = (((Angerable) entity).getTarget() != null) == IS_ANGRY;
-//                System.out.println("target="+((Angerable) entity).getTarget() );
-//            }else if (entity instanceof HostileEntity) {
-//                wasEntityTestedByAnUpdatableProperty = true;
-//                doesEntityMeetThisCaseTest = (((HostileEntity) entity).getTarget() != null) == IS_ANGRY;
-//            }else{
-//                doesEntityMeetThisCaseTest =false;
-//            }
-//        }
+        if (doesEntityMeetThisCaseTest && IS_ANGRY != null) {
+            if (entity instanceof EndermanEntity) {
+                wasEntityTestedByAnUpdatableProperty = true;
+                doesEntityMeetThisCaseTest = ((EndermanEntity) entity).isAngry() == IS_ANGRY;
+            }else if (entity instanceof BlazeEntity) {
+                wasEntityTestedByAnUpdatableProperty = true;
+                doesEntityMeetThisCaseTest = entity.isOnFire() == IS_ANGRY;
+            }else if (entity instanceof GuardianEntity) {
+                wasEntityTestedByAnUpdatableProperty = true;
+                doesEntityMeetThisCaseTest = (((GuardianEntity)entity).getBeamTarget() != null) == IS_ANGRY;
+            }else if (entity instanceof VindicatorEntity) {
+                wasEntityTestedByAnUpdatableProperty = true;
+                doesEntityMeetThisCaseTest = (((VindicatorEntity)entity).isAttacking()) == IS_ANGRY;
+            }else if (entity instanceof SpellcastingIllagerEntity) {
+                wasEntityTestedByAnUpdatableProperty = true;
+                doesEntityMeetThisCaseTest = (((SpellcastingIllagerEntity)entity).isSpellcasting()) == IS_ANGRY;
+            }else{
+                doesEntityMeetThisCaseTest =false;
+            }
+
+        }
 //        if (doesEntityMeetThisCaseTest && IS_ANGRY_WITH_CLIENT != null && MinecraftClient.getInstance().player != null) {
 //            wasEntityTestedByAnUpdatableProperty = true;
 //            if (entity instanceof HostileEntity) {
@@ -805,12 +813,12 @@ public class ETFTexturePropertyCase {
             wasEntityTestedByAnUpdatableProperty = true;
             doesEntityMeetThisCaseTest = ((CreeperEntity) entity).shouldRenderOverlay() == CREEPER_CHARGED;
         }
-        if (doesEntityMeetThisCaseTest && STATUS_EFFECT != null) {
+        if (doesEntityMeetThisCaseTest && entity instanceof LivingEntity && STATUS_EFFECT != null) {
             wasEntityTestedByAnUpdatableProperty = true;
             boolean found = false;
             for (StatusEffect effect :
                     STATUS_EFFECT) {
-                if (entity.hasStatusEffect(effect)) {
+                if (((LivingEntity)entity).hasStatusEffect(effect)) {
                     found = true;
                     break;
                 }
@@ -851,7 +859,7 @@ public class ETFTexturePropertyCase {
     }
 
     @NotNull
-    private ObjectImmutableList<String> readAllSpawnConditionsForCache(@NotNull LivingEntity entity) {
+    private ObjectImmutableList<String> readAllSpawnConditionsForCache(@NotNull Entity entity) {
         //check to speed up processing time
 
         //must be 6 length